Ability to direct to Single or List
